Code cleanup 2
