+++ Controller/CombatController.php
@@ -5,8 +5,6 @@
     use Model\Services\EnemyPartyService;
     use Model\Services\PlayerPartyService;
     use Model\Objects\CombatInfo;
-    use Model\Objects\EnemyParty;
-    use Model\Objects\PlayerParty;
 
     define('MAX_PARTY_MEMBERS_COUNT', 4);
 
@@ -21,8 +19,12 @@
             $playerParty = $playerPartyService->getPartyByName($playerPartyName);
             $enemyParty = $enemyPartyService->getEnemyPartyByName($enemyPartyName);
 
-            if($playerParty['success']) $playerParty = $playerParty['party'];
-            if($enemyParty['success']) $enemyParty = $enemyParty['party'];
+            if($playerParty['success']){
+                $playerParty = $playerParty['party'];
+            }
+            if($enemyParty['success']){
+                $enemyParty = $enemyParty['party'];
+            }
 
             $combatInfo = new CombatInfo($playerParty, $enemyParty, $this->generateTurns(), 0);
 
@@ -55,75 +57,6 @@
             session_write_close();
         }
 
-        public function combatBattle(){
-            session_start();
-
-            $targetId = (int)$_POST['targetId'];
-            $attackerId = (int)$_POST['attackerId'];
-            $combatInfo = json_decode($_POST['combatInfo'], true);
-
-            $playerParty = PlayerParty::jsonCreatePlayerParty($combatInfo['playerParty']);
-            $enemyParty = EnemyParty::jsonCreateEnemyParty($combatInfo['enemyParty']);
-            $turns = $combatInfo['turns'];
-            $currentTurn = $combatInfo['currentTurn'];
-
-            // player attack
-            $characterDamage = $playerParty->members[$attackerId]->getCharacterAttackDamage();
-            $enemyParty->members[$targetId]->takeDamage($characterDamage);
-
-            // check if enemy party is dead
-            if($this->checkParty($enemyParty)){
-                View::render('combatWin');
-                session_write_close();
-
-                return;
-            }
-
-            // enemy attack
-            $enemyTurn = $turns[$currentTurn][1];
-            if(!$enemyParty->members[$enemyTurn]->isCharacterDead()){
-                $enemyDamage = $enemyParty->members[$enemyTurn]->getCharacterAttackDamage();
-                $playerParty->members[rand(0, 3)]->takeDamage($enemyDamage);
-            }
-
-            // check if player party is dead
-            if($this->checkParty($playerParty)){
-                View::render('combatLose');
-                session_write_close();
-
-                return;
-            }
-
-            while(TRUE){
-                if($currentTurn >= 6) $currentTurn = 0;
-                else $currentTurn += 2;
-                $character = $playerParty->members[$turns[$currentTurn][1]];
-                if(!$character->isCharacterDead()) break;
-                else{
-                    $enemyTurn = $turns[$currentTurn][1];
-                    if(!$enemyParty->members[$enemyTurn]->isCharacterDead()){
-                        $enemyDamage = $enemyParty->members[$enemyTurn]->getCharacterAttackDamage();
-                        $playerParty->members[rand(0, 3)]->takeDamage($enemyDamage);
-                    }
-                }
-                if($this->checkParty($playerParty)){
-                    View::render('combatLose');
-                    session_write_close();
-    
-                    return;
-                }
-            };
-
-            $combatInfo = new CombatInfo($playerParty, $enemyParty, $turns, $currentTurn);
-            
-            $_SESSION['combatInfo'] = json_encode($combatInfo);
-
-            View::render('combatInfo');
-            session_write_close();
-
-            return;
-        }
-
         public function battle(){
             session_start();
 
@@ -131,8 +64,8 @@
             $attackerId = (int)$_POST['attackerId'];
             $combatInfo = json_decode($_POST['combatInfo'], true);
 
-            $playerParty = PlayerParty::jsonCreatePlayerParty($combatInfo['playerParty']);
-            $enemyParty = EnemyParty::jsonCreateEnemyParty($combatInfo['enemyParty']);
+            $playerParty = PlayerPartyService::jsonCreatePlayerParty($combatInfo['playerParty']);
+            $enemyParty = EnemyPartyService::jsonCreateEnemyParty($combatInfo['enemyParty']);
             $turns = $combatInfo['turns'];
             $currentTurn = $combatInfo['currentTurn'];
 
@@ -166,10 +99,18 @@
             }
 
             while(TRUE){
-                if($currentTurn >= 6) $currentTurn = 0;
-                else $currentTurn += 2;
+                if($currentTurn >= 6){
+                    $currentTurn = 0;
+                }
+                else{
+                    $currentTurn += 2;
+                }
+
                 $character = $playerParty->members[$turns[$currentTurn][1]];
-                if(!$character->isCharacterDead()) break;
+                
+                if(!$character->isCharacterDead()){
+                    break;
+                }
                 else{
                     $enemyTurn = $turns[$currentTurn][1];
                     if(!$enemyParty->members[$enemyTurn]->isCharacterDead()){
@@ -199,10 +140,14 @@
         private function checkParty($party){
             $deadCharacters = 0;
             for($i = 0;$i < MAX_PARTY_MEMBERS_COUNT;$i++){
-                if($party->members[$i]->isCharacterDead()) $deadCharacters++;
+                if($party->members[$i]->isCharacterDead()){
+                    $deadCharacters++;
+                }
             }
 
-            if($deadCharacters >= MAX_PARTY_MEMBERS_COUNT) return true;
+            if($deadCharacters >= MAX_PARTY_MEMBERS_COUNT){
+                return true;
+            }
             return false;
         }
     }
