+++ Model/Objects/Character.php
@@ -32,18 +32,6 @@
             return $character;
         }
 
-        public static function jsonCreateCharacter($json_decoded){
-            $character = new Character(
-                $json_decoded['characterName'],
-                $json_decoded['characterHealth'],
-                $json_decoded['characterAttackDamage'],
-                $json_decoded['characterMana']
-            );
-            $character->setCharacterId($json_decoded['characterId']);
-
-            return $character;
-        }
-
         public function getCharacterName(){
             return $this->characterName;
         }
