+++ Model/Objects/EnemyParty.php
@@ -23,21 +23,6 @@
             return $enemyParty;
         }
 
-        public static function jsonCreateEnemyParty($json_decoded){
-            $enemyPartyId = $json_decoded['enemyPartyId'];
-            $enemyPartyName = $json_decoded['enemyPartyName'];
-
-            $enemyParty = new EnemyParty($enemyPartyName, $enemyPartyId);
-            
-            foreach($json_decoded['members'] as $member){
-                $character = Character::jsonCreateCharacter($member);
-
-                $enemyParty->addMemberToEnemyParty($character);
-            }
-
-            return $enemyParty;
-        }
-
         public function getEnemyPartyId(){
             return $this->enemyPartyId;
         }
