+++ Model/Services/PlayerPartyService.php
@@ -34,6 +34,21 @@
             return $result;
         }
 
+        public static function jsonCreatePlayerParty($json_decoded){
+            $playerPartyId = $json_decoded['playerPartyId'];
+            $playerPartyName = $json_decoded['playerPartyName'];
+
+            $playerParty = new PlayerParty($playerPartyName, $playerPartyId);
+            
+            foreach($json_decoded['members'] as $member){
+                $character = CharacterService::jsonCreateCharacter($member);
+
+                $playerParty->addMemberToParty($character);
+            }
+
+            return $playerParty;
+        }
+
         public function getPartyByName($partyName)
         {
             $result = [
